+++ NEWS.md
@@ -5,6 +5,7 @@
 * Add class 'list' to `gar_discovery_*()` objects so they work with `jsonlite`
 * Add `gar_service_list()` and `gar_service_get()`
 * If using `gar_service_create()` and it already exists, will return existing via `gar_service_get()`
+* Move the setup wizard functions from `googleCloudRunner` to `googleAuthR` so they are available for all packages.
 
 # googleAuthR 1.2.1